full worked parser

+++ db/migrate/20170217225233__add_details.rb
@@ -1,8 +1,8 @@
 class AddDetails < ActiveRecord::Migration[5.0]
   def change
     add_column :ingredients, :calories, :integer
-    add_column :ingredients, :protein, :float
-    add_column :ingredients, :fat, :float
-    add_column :ingredients, :carbohydrate, :float
+    add_column :ingredients, :protein, :integer
+    add_column :ingredients, :fat, :integer
+    add_column :ingredients, :carbohydrate, :integer
   end
 end
